:art: refactor: #73 컴포넌트 폴더 구조 개선 및 책임 분리
