+++ src/pages/schedule/IdolDropdownPanel.tsx
